+++ wxc_sdk/__init__.py
@@ -25,6 +25,7 @@ from .reports import ReportsApi
 from .rest import RestSession
 from .room_tabs import RoomTabsApi
 from .rooms import RoomsApi
+from .status import StatusAPI
 from .team_memberships import TeamMembershipsApi
 from .teams import TeamsApi
 from .telephony import TelephonyApi
@@ -87,6 +88,8 @@ class WebexSimpleApi:
     rooms: RoomsApi
     #: Room tabs API :class:`room_tabs.RoomTabsApi`
     room_tabs: RoomTabsApi
+    #: Webex Status API :class:`status.StatusAPI`
+    status: StatusAPI
     #: Teams API :class:`teams.TeamsApi`
     teams: TeamsApi
     #: Team memberships API :class:`TeamMembershipsApi`
@@ -143,6 +146,7 @@ class WebexSimpleApi:
         self.reports = ReportsApi(session=session)
         self.rooms = RoomsApi(session=session)
         self.room_tabs = RoomTabsApi(session=session)
+        self.status = StatusAPI(session=session)
         self.teams = TeamsApi(session=session)
         self.team_memberships = TeamMembershipsApi(session=session)
         self.telephony = TelephonyApi(session=session)